completed experience section; revised projects, about, and css; deleted capybara gif and jobs file
