+++ src/styles/About.css
@@ -56,7 +56,9 @@ a.about-school:hover, a.about-major:hover {
 }
   
 .about-image {
-    padding-left: 40px;
+    padding-left: 170px;
+    padding-right: 100px;
+    padding-top: 100px;
 }
   
 .about-image img {
