+++ src/styles/Experience.css
@@ -1,6 +1,7 @@
 #experience {
     padding-left: 15%;
     padding-right: 25%;
+    padding-top: 10%;
     min-height: 60vh;
 }
 
@@ -13,3 +14,83 @@
     font-size: 46px;
     font-weight: 700;
 }
+
+.jobs-grid {
+    display: grid;
+    grid-template-columns: repeat(3, minmax(300px, 1fr));
+    grid-gap: 35px;
+    align-items: center;
+}
+
+.jobs-card {
+    position: relative;
+    cursor: default;
+    display: flex;
+    flex-direction: column;
+    height: 85%;
+    width: 90%;
+    padding: 2rem 1.75rem;
+    background-color: var(--beige);
+    border-radius: 18px;
+    box-shadow: 6px 10px 16px -8px rgba(0, 0, 0, 0.46);
+}
+
+.jobs-card:hover {
+    transform: translateY(-5px);
+    background-color: var(--slate);
+}
+
+.card-header {
+    margin-top: -25px;
+    display: flex;
+    padding: 1.25em 0 !important;
+    flex-direction: row;
+    justify-content: space-between;
+    border-bottom: none !important;
+    align-items: center;
+}
+
+.card-company {
+    font-style: italic;
+}
+
+.card-dates {
+    color: var(--grey);
+    font-size: 14px;
+}
+
+.card-title {
+    color: var(--black);
+    font-size: 22px;
+    font-weight: bold;
+}
+
+.card-desc {
+    padding-bottom: 3vh;
+}
+
+.card-desc ul {
+    list-style-type: none;
+    display: block;
+    padding: 0px;
+}
+
+.card-desc li {
+    position: relative;
+    padding-left: 0px;
+    padding-bottom: 10px;
+    font-size: 16px;
+    color: var(--grey);
+}
+
+.card-desc li::before {
+    content: "▹    ";
+    color: var(--dark-green);
+    position: relative;
+    left: 0;
+}
+
+li::before {
+    font-size: 20px;
+}
+
