--- src/styles/Jobs.css
@@ -1,119 +1,0 @@
-body {
-    /* background-color: #deeeff; */
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    height: 100vh
-}
-
-/* .container {
-    width: 600px;
-    padding: 30px;
-    background-color: white;
-    box-shadow: 0 2px 16px rgba(0, 0, 0, .1);
-    border-radius: 20px;
-}
-
-.tab_box {
-    width: 100%;
-    display: flex;
-    justify-content: space-around;
-    align-items: center;
-    border-bottom: 2px solid rgba(229, 229, 229);
-    position: relative;
-}
-
-.tab_box .tab_btn {
-    font-size: 18px;
-    font-weight: 600;
-    color: #919191;
-    background-color: none;
-    border: none;
-    padding: 18px;
-    cursor: pointer;
-}
-
-.tab_box .tab_btn.active {
-    color: #7360ff;
-}
-
-.content_box {
-    padding: 20px;
-}
-
-.content_box .content {
-    display: none;
-    animation: moving .5s ease;
-}
-
-@keyframes moving {
-    from{transform: translateX(50px); opacity: 0;}
-    to{transform: translateX(0px); opacity: 1;}
-}
-
-.content_box .content.active {
-    display: block;
-}
-
-.content_box .content h2 {
-    margin-bottom: 10px;
-}
-
-.line {
-    position: absolute;
-    top: 62px;
-    left: 17px;
-    width: 90px;
-    height: 5px;
-    background-color: #7360ff;
-    transition: all .3s ease-in-out;
-} */
-
-/* Basic styling for tabs and tab details */
-
-.tabs-container {
-    display: flex;
-}
-
-.tab-content {
-    width: 200px; /* Set your desired width for the tabs */
-    border-right: 1px solid #ccc; /* Border between tabs and tab details */
-}
-
-.tab {
-    padding: 10px;
-    cursor: pointer;
-    transition: background-color 0.3s ease;
-}
-
-.tab:hover {
-    background-color: #f0f0f0;
-}
-
-.tab.active {
-    background-color: #e0e0e0; /* Background color for the active tab */
-}
-
-.tab-details {
-    padding: 10px;
-}
-
-.tab-detail {
-    display: none;
-}
-
-.active-tab-detail {
-    display: block;
-    /* Add your animation styles for showing/hiding tab details */
-}
-
-/* Example animation styles (you can customize these) */
-@keyframes fadeInOut {
-    0% { opacity: 0; }
-    50% { opacity: 1; }
-    100% { opacity: 0; }
-}
-
-.active-tab-detail {
-    animation: fadeInOut 1s ease infinite; /* Example animation */
-}
