+++ src/styles/Projects.css
@@ -1,6 +1,7 @@
 #projects {
     padding-left: 15%;
     padding-right: 25%;
+    padding-top: 10%;
     min-height: 60vh;
 }
